CS-VoiceAgent/second/Workers/agents/pyproject.toml
[tool.uv]
constraint-dependencies = ['onnxruntime<1.24; python_version < "3.11"']

[tool.uv.sources]
livekit-agents = { workspace = true }
livekit-plugins-anam = { workspace = true }
livekit-plugins-anthropic = { workspace = true }
livekit-plugins-assemblyai = { workspace = true }
livekit-plugins-avatario = { workspace = true }
livekit-plugins-asyncai = { workspace = true }
livekit-plugins-aws = { workspace = true }
livekit-plugins-azure = { workspace = true }
livekit-plugins-baseten = { workspace = true }
livekit-plugins-bey = { workspace = true }
livekit-plugins-bithuman = { workspace = true }
livekit-plugins-cambai = { workspace = true }
livekit-plugins-cartesia = { workspace = true }
livekit-plugins-cerebras = { workspace = true }
livekit-plugins-clova = { workspace = true }
livekit-plugins-deepgram = { workspace = true }
livekit-plugins-elevenlabs = { workspace = true }
livekit-plugins-fal = { workspace = true }
livekit-plugins-fireworksai = { workspace = true }
livekit-plugins-fishaudio = { workspace = true }
livekit-plugins-gladia = { workspace = true }
livekit-plugins-google = { workspace = true }
livekit-plugins-groq = { workspace = true }
livekit-plugins-hedra = { workspace = true }
livekit-plugins-hamming = { workspace = true }
livekit-plugins-hume = { workspace = true }
livekit-plugins-inworld = { workspace = true }
livekit-plugins-krisp = { workspace = true }
livekit-plugins-keyframe = { workspace = true }
livekit-plugins-langchain = { workspace = true }
livekit-plugins-lemonslice = { workspace = true}
livekit-plugins-liveavatar = { workspace = true }
livekit-plugins-lmnt = { workspace = true }
livekit-plugins-minimax-ai = { workspace = true }
livekit-plugins-mistralai = { workspace = true }
livekit-plugins-murf = { workspace = true }
livekit-plugins-neuphonic = { workspace = true }
livekit-plugins-nltk = { workspace = true }
livekit-plugins-nvidia = { workspace = true }
livekit-plugins-openai = { workspace = true }
livekit-plugins-resemble = { workspace = true }
livekit-plugins-rime = { workspace = true }
livekit-plugins-runway = { workspace = true }
livekit-plugins-rtzr = { workspace = true }
livekit-plugins-sarvam = { workspace = true }
livekit-plugins-silero = { workspace = true }
livekit-plugins-simli = { workspace = true }
livekit-plugins-avatartalk = { workspace = true }
livekit-plugins-smallestai = { workspace = true }
livekit-plugins-simplismart = { workspace = true }
livekit-plugins-soniox = { workspace = true }
livekit-plugins-speechify = { workspace = true }
livekit-plugins-browser = { workspace = true }
livekit-plugins-speechmatics = { workspace = true }
livekit-plugins-spitch = { workspace = true }
livekit-plugins-tavus = { workspace = true }
livekit-plugins-telnyx = { workspace = true }
livekit-plugins-trugen = { workspace = true }
livekit-plugins-turn-detector = { workspace = true }
livekit-plugins-ultravox = { workspace = true }
livekit-plugins-upliftai = { workspace = true }
livekit-plugins-gradium = { workspace = true }
livekit-plugins-xai = { workspace = true }
livekit-plugins-phonic = { workspace = true }
livekit-plugins-did = { workspace = true }

[tool.uv.workspace]
members = ["livekit-plugins/*", "livekit-agents"]
exclude = [
    "livekit-plugins/community",
    "livekit-plugins/livekit-blingfire",
    "livekit-plugins/livekit-blockguard",
    "livekit-plugins/livekit-durable",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
    "mypy",
    "scikit-learn",
    "pytest",
    "ruff",
    "pytest-asyncio>=0.25.3",
    "jiwer>=3.1.0",
    "scipy>=1.13.1",
    "tiktoken>=0.9.0",
    "pytest-xdist>=3.8.0",
    "playwright>=1.58.0",
]
# Cybrix voice agent (examples/voice_agents/cybrix_dashscope_voice_agent.py):
# работает через inference-gateway, поэтому SDK провайдеров (dashscope, …) тут не нужны.
# razdel — для StreamingSentencizer (TTS режим commit, см. custom-plugins/streaming_sentencizer.py).
cybrix = [
    "razdel>=0.5.0",
]
docs = ["pdoc3", "setuptools", "beautifulsoup4", "markdownify"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [".github"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["livekit"]

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = ["--import-mode=importlib", "--ignore=examples"]
pythonpath = ["."]


[tool.mypy]
strict = true
disallow_any_generics = false
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "mcp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.genai"
follow_imports = "normal"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "boto3"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "aws_sdk_bedrock_runtime.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["aws_sdk_transcribe_streaming.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "smithy_aws_core.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "smithy_core.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "smithy_http.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "smithy_aws_event_stream.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mistralai.*"
follow_imports = "skip"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "speechmatics.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "livekit.plugins"
follow_untyped_imports = true
